--- cmd/lncli/cmd_reset_mission_control.go
@@ -1,26 +1,0 @@
-package main
-
-import (
-	"github.com/lightningnetwork/lnd/lnrpc/routerrpc"
-
-	"github.com/urfave/cli"
-)
-
-var resetMissionControlCommand = cli.Command{
-	Name:     "resetmc",
-	Category: "Mission Control",
-	Usage:    "Reset internal mission control state.",
-	Action:   actionDecorator(resetMissionControl),
-}
-
-func resetMissionControl(ctx *cli.Context) error {
-	ctxc := getContext()
-	conn := getClientConn(ctx, false)
-	defer conn.Close()
-
-	client := routerrpc.NewRouterClient(conn)
-
-	req := &routerrpc.ResetMissionControlRequest{}
-	_, err := client.ResetMissionControl(ctxc, req)
-	return err
-}
